RR-48 refactor camera logic
