+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/opmodes/MainTeleOp.java
@@ -12,7 +12,7 @@ import com.acmerobotics.relicrecovery.mech.GlyphLift;
 import com.acmerobotics.relicrecovery.mech.Periscope;
 import com.acmerobotics.relicrecovery.mech.RelicRecoverer;
 import com.acmerobotics.relicrecovery.util.LoggingUtil;
-import com.acmerobotics.relicrecovery.vision.VisionCamera;
+import com.acmerobotics.relicrecovery.vision.VuforiaCamera;
 import com.acmerobotics.relicrecovery.vision.VisionConstants;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
@@ -37,7 +37,7 @@ public class MainTeleOp extends OpMode {
 
     private boolean halfSpeed, secondControllerGlyph = true, changingMode;
 
-    private VisionCamera camera;
+    private VuforiaCamera camera;
 
     @Override
     public void init() {
@@ -52,9 +52,8 @@ public class MainTeleOp extends OpMode {
         Telemetry subsystemTelemetry = new MultipleTelemetry(loggingTelemetry, dashboard.getTelemetry());
         Telemetry allTelemetry = new MultipleTelemetry(telemetry, loggingTelemetry, dashboard.getTelemetry());
 
-        camera = new VisionCamera();
-        camera.setImageDir(LoggingUtil.getImageDir(this));
-        camera.initialize(VisionConstants.VUFORIA_PARAMETERS);
+        camera = new VuforiaCamera(VisionConstants.VUFORIA_PARAMETERS);
+        camera.initialize();
 
         drive = new MecanumDrive(hardwareMap, subsystemTelemetry, initialPose);
         frontLift = new GlyphLift(hardwareMap, subsystemTelemetry, GlyphLift.Side.FRONT);
