+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/opmodes/vision/CryptoboxVision.java
@@ -1,11 +1,10 @@
 package com.acmerobotics.relicrecovery.opmodes.vision;
 
 import com.acmerobotics.library.configuration.AllianceColor;
-import com.acmerobotics.relicrecovery.util.LoggingUtil;
 import com.acmerobotics.relicrecovery.vision.CryptoboxTracker;
 import com.acmerobotics.relicrecovery.vision.FpsTracker;
-import com.acmerobotics.relicrecovery.vision.VisionCamera;
 import com.acmerobotics.relicrecovery.vision.VisionConstants;
+import com.acmerobotics.relicrecovery.vision.VuforiaCamera;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
@@ -15,19 +14,18 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 @TeleOp(name = "Cryptobox Vision", group = "vision")
 public class CryptoboxVision extends OpMode {
-    private VisionCamera camera;
+    private VuforiaCamera camera;
     private CryptoboxTracker cryptoboxTracker;
     private FpsTracker fpsTracker;
 
     @Override
     public void init() {
-        camera = new VisionCamera();
+        camera = new VuforiaCamera(VisionConstants.VUFORIA_PARAMETERS);
         cryptoboxTracker = new CryptoboxTracker(AllianceColor.BLUE);
         fpsTracker = new FpsTracker();
         camera.addTracker(cryptoboxTracker);
         camera.addTracker(fpsTracker);
-        camera.setImageDir(LoggingUtil.getImageDir(this));
-        camera.initialize(VisionConstants.VUFORIA_PARAMETERS);
+        camera.initialize();
     }
 
     @Override
